requirements.txt
# Google ADK and AI
google-adk>=0.1.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Utilities
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# NLP tools
textstat>=0.7.3  # Readability scoring
nltk>=3.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0